fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38202 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/c64cart.c
@@ -856,7 +856,7 @@ static void cart_freeze_alarm_triggered(CLOCK offset, void *data)
 */
 void cartridge_trigger_freeze(void)
 {
-    int delay = lib_unsigned_rand(1, machine_get_cycles_per_frame());
+    int delay = lib_unsigned_rand(1, (unsigned int)machine_get_cycles_per_frame());
 
     cart_freeze_alarm_time = maincpu_clk + delay;
     alarm_set(cartridge_freeze_alarm, cart_freeze_alarm_time);
